gateway/requirements.txt
fastapi==0.58.0
uvicorn==0.11.5
PyJWT==1.7.1

# async http requests & fast dns resolving
aiohttp==3.6.2
aiodns==2.0.0

# for test purposes
flake8==3.8.3
ipdb==0.13.2
ipython==7.15.0
